Move JSON class to `io` module.
Close #6
